Fix testing bug due to document dependency
In order to make the store accessible accross js modules, the store was
initialized as soon as the script file was done loading. This created
a dependency on the document object, resulting in the mocha test
environment preventing me from executing the tests in the suite.

To resolve this, the store was initialized in the global execution
context rather than relying on timing the initialization just right.
The data that was being relied upon in the tested function is now being
passed in rather than being accessed in the function scope. This frees
up any dependencies on the test suite and allows for cross module access
to the store.