Move store to trips.js and create two separate stores
In this commit, I moved the store and data processing into the trips.js
file since that is the only place it is needed, so there is no sense in
trying to get it to communicate with scripts.js. Trips.js is the controller
for the trips page.

Scripts.js now has its own store called the userStore. This will store
the state of the current user and will be referred to when user data
is needed.