Parse user id into integer when getting value from localStorage
Since data is stored in localStorage as a string, the string format of
the user id was being passed to the backend which was breaking the filter
method and was not displaying any pending trips. Now that the user id
is being parsed into an integer, that bug is resolved.